+++ src/utils/points.js
@@ -5,7 +5,6 @@ const DateFormat = {
   TIME: 'HH:mm',
 };
 
-const currentDate = dayjs();
 
 function formatDateShort(date) {
   return date ? dayjs(date).format(DateFormat.DATE) : '';
@@ -16,11 +15,11 @@ function formatTime(date) {
 }
 
 function isFuturePoint(point) {
-  return dayjs(point.date_from) >= currentDate;
+  return dayjs(point.dateFrom) > dayjs();
 }
 
 function sortByTime(points) {
-  return points.sort(( a, b ) => (dayjs(b.date_from).isAfter(dayjs(a.date_from)) ? 1 : -1));
+  return points.sort(( a, b ) => (dayjs(b.dateFrom).isAfter(dayjs(a.dateFrom)) ? 1 : -1));
 }
 
 function sortByPrice(points) {
