+++ src/utils/sort.js
@@ -1,12 +1,22 @@
 import { SortType } from '../const.js';
 import { sortByTime, sortByPrice } from './points.js';
 
-const sort = {
-  [SortType.DAY]: (points) => sortByTime(points),
+const sortPointsToType = {
+  [SortType.DAY]: sortByTime,
   [SortType.EVENT]: (points) => points,
   [SortType.TIME]: (points) => points,
-  [SortType.PRICE]: (points) => sortByPrice(points),
+  [SortType.PRICE]: sortByPrice,
   [SortType.OFFER]: (points) => points,
 };
 
-export { sort };
+
+function generateSortItems(points) {
+  return Object.entries(sortPointsToType).map(
+    ([sortName, sortPoints]) => ({
+      name: sortName,
+      count: sortPoints(points).length,
+    }),
+  );
+}
+
+export { generateSortItems };
